refactor: colocate route-specific components
