+++ app/feed/feed-page-client.tsx
@@ -1,7 +1,7 @@
 "use client";
 import * as React from "react";
 import { Separator } from "@/components/ui/separator";
-import { TopBar } from "@/components/navigation";
+import { TopBar } from "./top-bar";
 import { FeedHeader, FeedSortBar, FeedList } from "@/components/feed";
 import { useSortedRides } from "@/hooks/useSortedRides";
 import { FeedPageClientProps } from "@/app/interface/main";
@@ -18,10 +18,13 @@ export default function FeedPageClient({
     <div className="bg-white min-h-screen">
       <FeedHeader feedbackUrl="https://forms.gle/DjypxU7tayRGVVMu5" />
       <div className="px-8">
-        <TopBar onResults={setRides} rides={localRides} />
+        <TopBar
+          onResults={setRides}
+          rides={localRides}
+        />
       </div>
       <div className="relative flex flex-1 flex-col p-6 bg-white">
-        <FeedSortBar sortBy={sortBy} setSortBy={setSortBy}/>
+        <FeedSortBar sortBy={sortBy} setSortBy={setSortBy} />
         <Separator className="mb-4" />
         <div className="pt-16 flex justify-center">
           <FeedList rides={sortedRides} bookmarkedRideIds={bookmarkedRideIds} />
