+++ components/navigation/index.ts
@@ -1,5 +1,4 @@
 export { NavProjects } from "./nav-projects";
 export { NavUser } from "./nav-user";
-export { TopBar } from "./top-bar";
 export { TopNavButtons } from "./top-nav-buttons";
 export { TeamSwitcher } from "./team-switcher";
