--- components/rides/index.ts
@@ -1,2 +1,0 @@
-export { EditRideDialog } from "./EditRideDialog";
-export { ShareYideDialog } from "./ShareYideDialog";
