+++ components/search/index.ts
@@ -1,2 +1,1 @@
 export { SearchBar } from "./search-bar";
-export { LocationCombobox } from "./location-combobox";
